requirements.txt
Flask==2.0.1
Flask-BasicAuth==0.2.0
flasgger==0.9.5
python-dotenv==0.19.2